refactor: clean code.
