+++ client/src/components/containers/CodeReviewMetricsContainers/CodeReviewFirstResponse/CodeReviewFirstResponseReviewersTable/codeReviewFirstResponseReviewersTableConstants.ts
@@ -1,0 +1,22 @@
+export const columns = [
+  {
+    id: "reviewer",
+    label: "Reviewer",
+    width: 220,
+  },
+  {
+    id: "addedTime",
+    label: "Added Time",
+    width: 185,
+  },
+  {
+    id: "votedTime",
+    label: "First Voted Time",
+    width: 185,
+  },
+  {
+    id: "firstReviewResponseTime",
+    label: "First Review Response Time",
+    width: 170,
+  },
+];
