+++ client/src/services/api/hooks/codeReview/codeReviewUtils.ts
@@ -78,7 +78,14 @@ export class CodeReviewMetricsUtils {
       const filteredThreads = this.#getFilteredThreads(pullRequest.threads, reviewersMap, pullRequest.authorId);
 
       return {
-        ...pullRequest,
+        id: pullRequest.id,
+        title: pullRequest.title,
+        status: pullRequest.status,
+        createdBy: pullRequest.createdBy,
+        authorId: pullRequest.authorId,
+        creationDate: pullRequest.creationDate,
+        closedDate: pullRequest.closedDate,
+        tags: pullRequest.tags,
         votes,
         votesHistory,
         votesTimeline: filteredVotesTimeline,
@@ -92,6 +99,8 @@ export class CodeReviewMetricsUtils {
           filteredVotesHistoryTimeline,
           reviewersMap,
         ),
+        mergeTimeInSeconds: pullRequest.mergeTimeInSeconds,
+        url: pullRequest.url,
       };
     });
